change mentsu-util.js to mentsu-util.ts
